testing done

+++ src/App/modules/user/user.service.ts
@@ -3,6 +3,9 @@ import bcrypt from "bcrypt";
 import { TUser } from "./user.interface";
 import config from "../../config";
 import { TJwtDecode } from "../../interface/global.type";
+import CustomError from "../../errors/customError";
+import { StatusCodes } from "http-status-codes";
+import { JwtPayload } from "jsonwebtoken";
 const prisma = new PrismaClient();
 
 export const createUserDB = async (payload: TUser) => {
@@ -62,8 +65,15 @@ export const getAUserDB = async (payload: TJwtDecode) => {
   return allUser;
 };
 
-export const updateUserDB = (user: TJwtDecode, payload: Partial<User>) => {
-  const update = prisma.user.update({
+export const updateUserDB = async (
+  user: TJwtDecode,
+  payload: Partial<User>
+) => {
+  await prisma.user.findUniqueOrThrow({
+    where: { email: user.email },
+  });
+
+  const update = await prisma.user.update({
     where: { email: user.email },
     data: payload,
     select: {
@@ -77,7 +87,16 @@ export const updateUserDB = (user: TJwtDecode, payload: Partial<User>) => {
   return update;
 };
 
-export const deleteUserDB = async (userId: string) => {
+export const deleteUserDB = async (userId: string, user: TJwtDecode) => {
+  const userInfo = await prisma.user.findUniqueOrThrow({
+    where: { id: userId },
+  });
+  if (userInfo.email === user.email) {
+    throw new CustomError(
+      StatusCodes.BAD_REQUEST,
+      "You can't modify yourself!"
+    );
+  }
   const deleteResult = await prisma.user.delete({
     where: { id: userId },
   });
@@ -85,14 +104,26 @@ export const deleteUserDB = async (userId: string) => {
   return deleteResult;
 };
 
-export const changeUserRoleAndStatusDB = async (payload: {
-  email?: string;
-  id?: string;
-  status?: userStatus;
-  role?: UserRole;
-}) => {
+export const changeUserRoleAndStatusDB = async (
+  payload: {
+    email?: string;
+    id?: string;
+    status?: userStatus;
+    role?: UserRole;
+  },
+  user: JwtPayload
+) => {
+  const userInfo = await prisma.user.findUniqueOrThrow({
+    where: { id: payload.id, email: payload.email },
+  });
+  if (userInfo.email === user.email) {
+    throw new CustomError(
+      StatusCodes.BAD_REQUEST,
+      "You can't modify yourself!"
+    );
+  }
+
   if (payload.status) {
-    console.log(payload, "y");
     return await prisma.user.update({
       where: { id: payload.id },
       data: { status: payload.status },
@@ -108,8 +139,6 @@ export const changeUserRoleAndStatusDB = async (payload: {
     },
   });
 
-  console.log(payload.email, "xsxss");
-  return;
   const result = await prisma.user.update({
     where: { email: payload.email },
     data: { role: payload.role },
